--- a/bin/anthology/bibtex.py
+++ b/bin/anthology/bibtex.py
@@ -30,7 +30,7 @@ def read_bibtex(bibfilename):
             bibstring = bibbytes.decode(encoding)
         except UnicodeDecodeError:
             continue
-        logging.debug("{}: using {} encoding".format(bibfilename, encoding))
+        logging.debug(f"{bibfilename}: using {encoding} encoding")
         break
     else:
         logging.warning("couldn't figure out encoding; using ascii with escapes")