--- a/bin/anthology/papers.py
+++ b/bin/anthology/papers.py
@@ -59,6 +59,7 @@ class Paper:
                 "meta_date",
                 "url",
                 "pdf",
+                "xml_url",
             ):
                 continue