This PR implements a bunch of changes to speed up instantiation of the Anthology class (re #835).

They're a mix of:
- delaying computations until the values are actually needed;
- general optimizations (using f-strings instead of `.format` gives a surprisingly large boost!);
- introducing a `fast_load` parameter to the Anthology class which skips some error checking and computes fewer information in the AnthologyIndex by default.

Here's a list of the most relevant commits and the effect on instantiation time as measured by `python -m timeit -n 2 -r 3 "from anthology import Anthology; Anthology(importdir='../data/')"`, additionally with `fast_load=True` from the commits that introduced it:

- 881e19c1 (master, before any changes)
  11.6 sec
- f9f8e3d1 Lazily infer URLs
  10.4 sec
- d3c69f67 f-strings everywhere
  8.46 sec
- e5a22878 Cache Anthology ID construction
  7.82 sec
- f9df8ab7 Replace re.sub with while-loop
  7.2 sec
- a39a7eef Refactor variant scoring and person ID sorting
  7.45 sec *(this one seems counterproductive but I still think it makes sense to refactor this; maybe I'll revisit why it makes things slower)*
- 13edbe94 Lazily render paper titles
  5.83 sec
- d3b7e980 Add fast_load option, disabling some error checking during load
  5.5 sec
- f9b88e15 Load id_to_used & coauthors lazily if fast_load=True
  4.34 sec

We should check carefully that nothing in the appearance/generation of the website changed, as I already had to make some fixes because of that.