@mbollmann 2021-08-05
Seeing some diffs regarding canonical names locally:

```
< <a href=/people/y/yujie-zhang/>Yujie Zhang</a>,
---
> <a href=/people/y/yujie-zhang/>Yujie Zhang (张玉洁)</a>,
```

...but strangely not on the preview branch. I'll turn this into a draft until I can better verify that nothing has functionally changed.