@mbollmann 2021-08-08
Better be more careful with this one, I still have the author name differences to investigate. This would be a really good situation to have some proper unit tests for, as right now I'm basically relying on building and diffing the full website, which is a bit clunky to say the least...