@mbollmann 2021-08-16
This should be ready for review now. The only differences in the build that I could see, compared to the master branch, are:

- The generated YAML files often use IDs and pointers instead of repeating people's names; this is likely a consequence of introducing caching for name lookups *(...and is a good thing)*.
- The order of name variants on people's pages is different, likely as it previously wasn't sorted, so not guaranteed to be in a specific order *(...but it now is)*.

It would be great if others could test this locally, especially with regards to the instantiation times on this branch compared to `master`. (I currently only have one device that I can't reliably test with, since it had some overheating issues and I'm throttling CPU power...)