@mbollmann 2021-08-28
So, the temporary fix for this [is really ugly](https://github.com/acl-org/acl-anthology/commit/16e838b6f0f44a5c92bfa4d6adbe779b2ba4887d). In hindsight, storing local name variants in `PersonName` objects was a mistake from a coding perspective, but one that I also didn't anticipate at that time. We shouldn't see that latest commit as anything but a band-aid solution until a proper refactoring of how names work, which I really want to prioritize now.